01_Magical_Shop/requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv